requirements.txt
grpcio
grpcio-tools
numpy
opencv-python
